// src/models/Product.ts
export interface Product {
  apiFeaturedImage: string;
  brand: string;
  category: string;
  description: string;
  id: string;
  name: string;
  price: string;
  priceSign: string;
  productID: number;
  productLink: string;
  productType: string;
  websiteLink: string;
  rating: number;
  tagList: string[];
}